test: Run package tests with verdaccio
