+++ .github/workflows/ci.yml
@@ -39,6 +39,22 @@ jobs:
           path: |
             lib*/**/*
 
+      # Publish new version to local registry
+      - name: Run local npm registry
+        run: ./scripts/local-registry.sh &
+
+      - name: Publish package locally
+        run: npm publish
+
+      - name: Upload local registry files
+        uses: actions/upload-artifact@v4.6.1
+        with:
+          name: local-registry
+          retention-days: 1
+          include-hidden-files: true
+          path: |
+            .verdaccio-storage*/**/*
+
   lint:
     runs-on: ubuntu-latest
     needs:
@@ -93,6 +109,14 @@ jobs:
         with:
           name: dist-files
 
+      - name: Download local registry files
+        uses: actions/download-artifact@v4.1.9
+        with:
+          name: local-registry
+
+      - name: Run local npm registry
+        run: ./scripts/local-registry.sh &
+
       - name: Test CLI
         run: npm run test:cli
 
@@ -125,6 +149,14 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: dist-files
+
+      - name: Download local registry files
+        uses: actions/download-artifact@v4.1.9
+        with:
+          name: local-registry
+
+      - name: Run local npm registry
+        run: ./scripts/local-registry.sh &
 
       - name: Test webpack-plugin
         run: npm run test:webpack-plugin -- -t ${{ matrix.webpack }}
