+++ package.json
@@ -33,16 +33,18 @@
     }
   },
   "scripts": {
-    "build": "rimraf ./lib && npx tsc -p ./tsconfig.esm.json && rollup -c && npm run build-type",
+    "build": "npm run clean && rollup -c && npm run build-type",
     "build-type": "echo '{\"type\":\"commonjs\"}' > lib/cjs/package.json && echo '{\"type\":\"module\"}' > lib/esm/package.json",
+    "clean": "rimraf ./lib",
+    "clean-deps": "rimraf ./test/cli/node_modules ./test/webpack/*/node_modules ./node_modules",
     "lint": "eslint .",
     "bump": "./scripts/bump.sh",
     "release": "./scripts/release.sh",
-    "pretest:webpack-plugin": "npm install webpack4@npm:webpack@^4.47.0",
-    "test": "npm run test:unit && npm run test:cli && npm run test:webpack-plugin",
-    "test:unit": "jest src",
-    "test:cli": "jest test/cli.test.js",
-    "test:webpack-plugin": "jest test/webpack-plugin",
+    "pretest:package-cli": "cd test/cli && npm install",
+    "pretest:package-webpack": "cd test/webpack/webpack4 && npm install --cwd ../webpack5 && npm install --cwd test/webpack/webpack5",
+    "test": "jest src",
+    "test:package-cli": "jest test/cli",
+    "test:webpack-plugin": "jest test/webpack",
     "prepublishOnly": "npm run build"
   },
   "engines": {
@@ -100,8 +102,8 @@
     "ts-jest": "29.2.6",
     "typescript": "5.8.2",
     "typescript-eslint": "8.26.0",
-    "webpack": "5.97.1",
-    "webpack4": "npm:webpack@^4.47.0"
+    "verdaccio": "5.26.2",
+    "webpack": "5.97.1"
   },
   "dependencies": {
     "@bundle-stats/plugin-webpack-filter": "4.19.0-beta.2",
