+++ test/cli/cli.test.js
@@ -1,5 +1,5 @@
 const { exec } = require('child_process');
-const { createServer } = require('./utils');
+const { createServer } = require('../utils');
 
 const MOCK_SERVER_PORT = 5998;
 
@@ -15,21 +15,21 @@ describe('CLI', () => {
   });
 
   test('should return error if config is missing', (done) => {
-    exec('./bin/index.js', (_, __, sterr) => {
+    exec('npx relative-ci', (_, __, sterr) => {
       expect(sterr).toContain('relativeci.config.js file is missing!');
       done();
     });
   });
 
   test('should return error if webpack stats is missing', (done) => {
-    exec('cd test/cli/missing-stats && ../../../bin/index.js', (_, __, sterr) => {
+    exec('cd test/cli/missing-stats && npx relative-ci', (_, __, sterr) => {
       expect(sterr).toContain('file does not exists');
       done();
     });
   });
 
   test('should return error if webpack stats data is invalid', (done) => {
-    exec('cd test/cli/invalid-data && ../../../bin/index.js', (_, __, sterr) => {
+    exec('cd test/cli/invalid-data && npx relative-ci', (_, __, sterr) => {
       expect(sterr).toContain('Invalid stats structure');
       done();
     });
@@ -39,7 +39,7 @@ describe('CLI', () => {
     exec(`cd test/cli/valid-data &&
       RELATIVE_CI_ENDPOINT=http://localhost:${MOCK_SERVER_PORT}/save \
       RELATIVE_CI_SLUG=org/project \
-      ../../../bin/index.js`, (_, __, sterr) => {
+      npx relative-ci`, (_, __, sterr) => {
       expect(sterr).toContain('parameter is missing');
       done();
     });
@@ -51,7 +51,7 @@ describe('CLI', () => {
         RELATIVE_CI_ENDPOINT=http://localhost:${MOCK_SERVER_PORT}/save \
         RELATIVE_CI_SLUG=org/project \
         RELATIVE_CI_KEY=abc123 \
-        ../../../bin/index.js
+        npx relative-ci
       `,
       (_, stdout, sterr) => {
         expect(sterr).toEqual('');
@@ -67,7 +67,7 @@ describe('CLI', () => {
         RELATIVE_CI_ENDPOINT=http://localhost:${MOCK_SERVER_PORT}/save \
         RELATIVE_CI_SLUG=org/project \
         RELATIVE_CI_KEY=abc123 \
-        ../../../bin/index.js --config-dir app
+        npx relative-ci --config-dir app
       `,
       (_, stdout, sterr) => {
         expect(sterr).toEqual('');
