--- test/webpack/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "webpack",
-  "version": "1.0.0",
-  "scripts": {
-    "build": "webpack-cli --mode=production --config webpack.config.js",
-    "build-esm": "cd config-module && npm run build",
-    "build-mjs": "webpack-cli --mode=production --config webpack.config.mjs",
-    "build-ts": "cross-env NODE_OPTIONS='--loader ts-node/esm' NODE_NO_WARNINGS=1 webpack-cli --mode=production --config webpack.config.ts",
-    "build-fail-on-error": "webpack-cli --mode=production --config webpack-fail-on-error.config.js"
-  },
-  "devDependencies": {
-    "@relative-ci/agent": "file://../../",
-    "cross-env": "^7.0.3",
-    "ts-node": "^10.0.0",
-    "webpack": "file://../../node_modules/webpack",
-    "webpack-cli": "^6.0.1"
-  }
-}
