+++ test/webpack/webpack4/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@relative-ci/agent-test-webpack-4",
+  "version": "1.0.0",
+  "devDependencies": {
+    "@relative-ci/agent": "*",
+    "memory-fs": "^0.5.0",
+    "webpack": "^4"
+  }
+}
