+++ test/webpack/webpack4/webpack-fail-on-error.config.js
@@ -1,0 +1,12 @@
+// eslint-disable-next-line
+const { RelativeCiAgentWebpackPlugin } = require('@relative-ci/agent');
+
+module.exports = {
+  context: __dirname,
+  plugins: [
+    new RelativeCiAgentWebpackPlugin({
+      includeCommitMessage: false,
+      failOnError: true,
+    }),
+  ],
+};
