+++ test/webpack/webpack4/webpack.config.js
@@ -1,0 +1,11 @@
+// eslint-disable-next-line
+const { RelativeCiAgentWebpackPlugin } = require('@relative-ci/agent');
+
+module.exports = {
+  context: __dirname,
+  plugins: [
+    new RelativeCiAgentWebpackPlugin({
+      includeCommitMessage: false,
+    }),
+  ],
+};
