+++ test/webpack/webpack5/config-module/package.json
