+++ test/webpack/webpack5/src/index.js
@@ -1,0 +1,1 @@
+console.log('Hello world!');
