+++ test/webpack/webpack5/webpack-plugin-build.test.js
@@ -1,12 +1,13 @@
 const { exec } = require('child_process');
-const { createServer } = require('./utils');
+
+const { createServer } = require('../../utils');
 
 const MOCK_SERVER_PORT = 5998;
 
 const testCases = [
   { type: 'commonjs', run: 'build' },
   { type: 'mjs', run: 'build-mjs' },
-  { type: 'esm', run: 'build-esm' },
+  { type: 'esm', cwd: 'config-module', run: 'build' },
   { type: 'ts', run: 'build-ts' },
 ];
 
@@ -23,7 +24,7 @@ describe('webpack-plugin / build / webpack5', () => {
 
   testCases.forEach((testCase) => {
     test(`should build successfully with webpack ${testCase.type} config`, (done) => {
-      exec(`cd test/webpack/${testCase.cwd || ''} &&
+      exec(`cd test/webpack/webpack5/${testCase.cwd || ''} &&
           CI=true \
           RELATIVE_CI_ENDPOINT=http://localhost:${MOCK_SERVER_PORT}/save \
           RELATIVE_CI_SLUG=org/project \
