+++ test/webpack/webpack5/webpack-plugin.test.js
@@ -1,19 +1,18 @@
 jest.mock('node-fetch');
-jest.mock('webpack', () => require('webpack4')); // eslint-disable-line
 
-const webpack = require('webpack4');
+const webpack = require('webpack');
 const MemoryFS = require('memory-fs');
+// eslint-disable-next-line
 const fetch = require('node-fetch');
 
-const { INGEST_MOCK } = require('./utils');
-const webpackStats = require('./__snapshots__/webpack-4-stats.json');
-const appConfig = require('./webpack/webpack.config');
-const appFailOnErrorConfig = require('./webpack/webpack-fail-on-error.config');
+const webpack5Stats = require('../../__snapshots__/webpack-5-stats.json');
 const {
-  ENV_DEFAULT, clearCustomEnv, getMockRequest, setCustomEnv,
-} = require('./utils');
+  ENV_DEFAULT, INGEST_MOCK, clearCustomEnv, getMockRequest, setCustomEnv,
+} = require('../../utils');
+const appConfig = require('./webpack.config');
+const appFailOnErrorConfig = require('./webpack-fail-on-error.config');
 
-describe('webpack-plugin / webpack4', () => {
+describe('webpack-plugin / webpack5', () => {
   afterEach(() => {
     clearCustomEnv();
     jest.clearAllMocks();
@@ -32,7 +31,7 @@ describe('webpack-plugin / webpack4', () => {
     compiler.outputFileSystem = new MemoryFS();
 
     compiler.run((error, stats) => {
-      expect(error).toEqual(null);
+      expect(error).toBeNull();
       expect(stats.hasErrors()).toBe(false);
       expect(fetch).toHaveBeenCalledTimes(1);
       expect(fetch).toHaveBeenCalledWith(
@@ -42,7 +41,7 @@ describe('webpack-plugin / webpack4', () => {
             webpack: {
               stats: {
                 hash: stats.hash,
-                ...webpackStats,
+                ...webpack5Stats,
               },
             },
           },
@@ -53,15 +52,24 @@ describe('webpack-plugin / webpack4', () => {
     });
   });
 
-  test('should warn, not ingest and not throw on params error', (done) => {
+  test('should warn, not ingest, and not throw on params error', (done) => {
     setCustomEnv({ RELATIVE_CI_KEY: '' });
 
     const compiler = webpack(appConfig);
     compiler.outputFileSystem = new MemoryFS();
 
+    const log = jest.spyOn(compiler, 'infrastructureLogger');
+
     compiler.run((error, stats) => {
       expect(stats.hasErrors()).toBe(false);
-      expect(error).toEqual(null);
+      expect(error).toBeNull();
+      expect(log).toHaveBeenLastCalledWith(
+        'RelativeCiAgent',
+        'warn',
+        expect.arrayContaining([
+          expect.objectContaining({ message: expect.stringContaining('"key" parameter is missing') }),
+        ]),
+      );
       expect(fetch).not.toHaveBeenCalled();
 
       done();
@@ -76,10 +84,19 @@ describe('webpack-plugin / webpack4', () => {
     const compiler = webpack(appConfig);
     compiler.outputFileSystem = new MemoryFS();
 
+    const log = jest.spyOn(compiler, 'infrastructureLogger');
+
     compiler.run((error, stats) => {
-      expect(error).toEqual(null);
+      expect(error).toBeNull();
       expect(stats.hasErrors()).toBe(false);
       expect(fetch).toHaveBeenCalledTimes(1);
+      expect(log).toHaveBeenLastCalledWith(
+        'RelativeCiAgent',
+        'warn',
+        expect.arrayContaining([
+          expect.objectContaining({ message: 'Error ingesting data!' }),
+        ]),
+      );
       done();
     });
   });
@@ -96,7 +113,7 @@ describe('webpack-plugin / webpack4', () => {
       compiler.run((error, stats) => {
         expect(error).toBeNull();
         expect(stats.hasErrors()).toBe(true);
-        expect(stats.toJson().errors[0]).toMatch(/Error ingesting data/);
+        expect(stats.toJson().errors[0]).toMatchObject({ message: /Error ingesting data/ });
         done();
       });
     } catch (err) {
