+++ test/webpack/webpack5/webpack.config.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line
-const { RelativeCiAgentWebpackPlugin } = require('../..');
+const { RelativeCiAgentWebpackPlugin } = require('@relative-ci/agent');
 
 module.exports = {
   mode: 'production',
@@ -7,7 +7,6 @@ module.exports = {
   plugins: [
     new RelativeCiAgentWebpackPlugin({
       includeCommitMessage: false,
-      failOnError: true,
     }),
   ],
 };
