+++ test/webpack/webpack5/webpack.config.ts
