+++ verdaccio.yaml
@@ -1,0 +1,19 @@
+storage: .verdaccio-storage
+packages:
+  '@relative-ci/agent':
+    access: $all
+    publish: $anonymous
+  '@*/*':
+    access: $all
+    publish: $all
+    proxy: npmjs
+  '**':
+    access: $all
+    publish: $all
+    proxy: npmjs
+publish:
+  allow_offline: true
+uplinks:
+  npmjs:
+    url: https://registry.npmjs.org/
+    max_fails: 100
